requirements.txt
streamlit>=1.31
pandas>=2.0
python-dateutil>=2.8
openai>=1.0.0
pydantic>=2.0
rapidfuzz>=3.0
langgraph>=0.2.0
pytest>=8.0
